+++ src/cfbl/cfrusb.f90
@@ -85,8 +85,9 @@ double precision bc_en(nfabor), bc_pr(nfabor), bc_vel(3,nfabor)
 integer          iel, l_size
 integer          ien
 
-double precision und, uni, rund, runi, cd, ci, cd2(1), ci2(1)
-double precision rrus, runb
+double precision inv_surfbn, rnx, rny, rnz
+double precision b_vel_n, c_vel_n, b_masfl, c_masfl, b_c, c_c, b_c2(1), c_c2(1)
+double precision rrus, r_b_masfl, r_b_vel_n
 double precision, dimension(:,:), pointer :: cofacv
 double precision, dimension(:), pointer :: coface
 double precision, dimension(:), pointer :: crom, brom
@@ -94,7 +95,7 @@ double precision, dimension(:,:), pointer :: vel
 double precision, dimension(:), pointer :: cvar_pr, cvar_en, cpro_cp, cpro_cv
 double precision, dimension(:), pointer :: rvoid
 
-double precision cpi(1), cvi(1)
+double precision c_cp(1), c_cv(1)
 
 !===============================================================================
 
@@ -124,48 +125,57 @@ iel   = ifabor(ifac)
 
 if (icp.ge.0) then
   call field_get_val_s(icp, cpro_cp)
-  cpi(1) = cpro_cp(iel)
+  c_cp(1) = cpro_cp(iel)
 else
   cpro_cp => rvoid1
-  cpi(1) = 0.d0
+  c_cp(1) = 0.d0
 endif
 
 if (icv.ge.0) then
   call field_get_val_s(icv, cpro_cv)
-  cvi(1) = cpro_cv(iel)
+  c_cv(1) = cpro_cv(iel)
 else
   cpro_cv => rvoid1
-  cvi(1) = 0.d0
+  c_cv(1) = 0.d0
 endif
 
 !===============================================================================
 ! 1. COMPUTE VALUES NEEDED FOR RUSANOV SCHEME
 !===============================================================================
 
-und   = (bc_vel(1,ifac)*surfbo(1,ifac)                          &
-       + bc_vel(2,ifac)*surfbo(2,ifac)                          &
-       + bc_vel(3,ifac)*surfbo(3,ifac))/surfbn(ifac)
-uni   = (vel(1,iel)*surfbo(1,ifac)                            &
-       + vel(2,iel)*surfbo(2,ifac)                            &
-       + vel(3,iel)*surfbo(3,ifac))/surfbn(ifac)
-rund  = brom(ifac)*und
-runi  = crom(iel)     *uni
+inv_surfbn = 1. / surfbn(ifac)
+rnx = surfbo(1,ifac) * inv_surfbn
+rny = surfbo(2,ifac) * inv_surfbn
+rnz = surfbo(3,ifac) * inv_surfbn
+
+b_vel_n = bc_vel(1,ifac)*rnx + bc_vel(2,ifac)*rny + bc_vel(3,ifac)*rnz
+c_vel_n = vel(1,iel)*rnx + vel(2,iel)*rny + vel(3,iel)*rnz
+b_masfl  = brom(ifac)*b_vel_n
+c_masfl  = crom(iel)*c_vel_n
 
 l_size = 1
 
-call cs_cf_thermo_c_square(cpi, cvi, &
+call cs_cf_thermo_c_square(c_cp, c_cv, &
                            bc_pr(ifac:ifac), brom(ifac:ifac), &
-                           rvoid, rvoid, rvoid, cd2, l_size)
-call cs_cf_thermo_c_square(cpi, cvi, &
+                           rvoid, rvoid, rvoid, b_c2, l_size)
+call cs_cf_thermo_c_square(c_cp, c_cv, &
                            cvar_pr(iel:iel), crom(iel:iel), &
-                           rvoid, rvoid, rvoid, ci2, l_size)
+                           rvoid, rvoid, rvoid, c_c2, l_size)
 
-cd    = sqrt(cd2(1))
-ci    = sqrt(ci2(1))
-rrus  = max(abs(und)+cd,abs(uni)+ci)
+b_c    = sqrt(b_c2(1))
+c_c    = sqrt(c_c2(1))
+rrus  = max(abs(b_vel_n)+b_c, abs(c_vel_n)+c_c)
 
-runb  = 0.5d0*(brom(ifac)*und+crom(iel)*uni)          &
-      - 0.5d0*rrus*(brom(ifac)-crom(iel))
+! boundary mass flux computed with Rusanov scheme
+r_b_masfl  = 0.5d0*(b_masfl+c_masfl) - 0.5d0*rrus*(brom(ifac)-crom(iel))
+
+! Rusanov normal velocity (computed using boundary density)
+r_b_vel_n = r_b_masfl / brom(ifac)
+
+! Update velocity boundary condition
+bc_vel(1,ifac) = bc_vel(1,ifac) + (r_b_vel_n - b_vel_n)*rnx
+bc_vel(2,ifac) = bc_vel(2,ifac) + (r_b_vel_n - b_vel_n)*rny
+bc_vel(3,ifac) = bc_vel(3,ifac) + (r_b_vel_n - b_vel_n)*rnz
 
 !===============================================================================
 ! 2. CONVECTIVE RUSANOV FLUX
@@ -179,29 +189,29 @@ icvfli(ifac) = 1
 ! Momentum flux (the centered pressure contribution is directly taken into account
 ! in the pressure BC)
 cofacv(1,ifac) = suffbn(ifac)*                                                  &
-                 0.5d0*( rund*bc_vel(1,ifac) + runi*vel(1,iel)                   &
-                         -rrus*(brom(ifac)*bc_vel(1,ifac)                        &
-                         -crom(iel)*vel(1,iel)) )
+                 0.5d0*( b_masfl*bc_vel(1,ifac) + c_masfl*vel(1,iel)            &
+                        -rrus*(brom(ifac)*bc_vel(1,ifac)                        &
+                        -crom(iel)*vel(1,iel)) )
 
 cofacv(2,ifac) = suffbn(ifac)*                                                  &
-                 0.5d0*( rund*bc_vel(2,ifac) + runi*vel(2,iel)                   &
-                         -rrus*( brom(ifac)*bc_vel(2,ifac)                       &
-                         -crom(iel)*vel(2,iel)) )
+                 0.5d0*( b_masfl*bc_vel(2,ifac) + c_masfl*vel(2,iel)            &
+                        -rrus*( brom(ifac)*bc_vel(2,ifac)                       &
+                        -crom(iel)*vel(2,iel)) )
 
 cofacv(3,ifac) = suffbn(ifac)*                                                  &
-                 0.5d0*( rund*bc_vel(3,ifac) + runi*vel(3,iel)                   &
-                         -rrus*(brom(ifac)*bc_vel(3,ifac)                        &
-                         -crom(iel)*vel(3,iel)) )
+                 0.5d0*( b_masfl*bc_vel(3,ifac) + c_masfl*vel(3,iel)            &
+                        -rrus*(brom(ifac)*bc_vel(3,ifac)                        &
+                        -crom(iel)*vel(3,iel)) )
 
 ! BC for the pressure gradient in the momentum balance
 bc_pr(ifac) = 0.5d0 * (bc_pr(ifac) + cvar_pr(iel))
 
 ! Total energy flux
 coface(ifac) = suffbn(ifac)*                                                    &
-               0.5d0*( rund*bc_en(ifac) + runi*cvar_en(iel)                  &
-                       +und*bc_pr(ifac) + uni*cvar_pr(iel)                   &
-                       -rrus*(brom(ifac)*bc_en(ifac)                         &
-                       -crom(iel)*cvar_en(iel)) )
+               0.5d0*( b_masfl*bc_en(ifac) + c_masfl*cvar_en(iel)               &
+                      +b_vel_n*bc_pr(ifac) + c_vel_n*cvar_pr(iel)               &
+                      -rrus*(brom(ifac)*bc_en(ifac)                             &
+                      -crom(iel)*cvar_en(iel)) )
 
 return